fix: remove unused imports
