+++ package.json
@@ -32,6 +32,7 @@
   },
   "homepage": "https://github.com/amangautam72/react-native-select-bottom-list#readme",
   "devDependencies": {
+    "@types/react": "^18.2.28",
     "metro-react-native-babel-preset": "^0.73.1",
     "typescript": "^4.8.4"
   },
